backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Firebase Admin SDK
firebase-admin==6.3.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.8

# Database & Cache
redis==5.0.1
hiredis==2.2.3

# Data processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# AI/ML integrations
openai==1.3.7
google-generativeai==0.3.2
anthropic==0.7.8

# Analytics
google-cloud-bigquery==3.13.0

# Background jobs
celery==5.3.4
redis-py-cluster==2.1.3

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
typer==0.9.0
rich==13.7.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
sentry-sdk==1.38.0